+++ src/main/java/com/roboskeletron/task13/interfaces/ITransform.java
@@ -5,8 +5,8 @@ import com.roboskeletron.task13.primitives.Vector2D;
 
 public interface ITransform {
     void translate(Vector2D vector);
-    void translateX(int x);
-    void translateY(int y);
+    void translateX(double x);
+    void translateY(double y);
     void setPosition(Point2D point);
-    void setPosition(int x, int y);
+    void setPosition(double x, double y);
 }
